backend/requirements.txt
# Google ADK and Gemini
google-adk>=1.0.0
google-genai>=1.0.0

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
